requirements.txt
defusedxml==0.5.0
Django==1.10.6
django-allauth==0.31.0
oauthlib==2.0.1
pkg-resources==0.0.0
psycopg2==2.7.1
python-openid==2.2.5
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
